# src/api/docs/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "TTHK Timetable API",
    "description": "This is the unofficial API to access TTHK's timetable data",
    "version": "1.0.0"
  },
  "host": "localhost:5637",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/timetable/{timetableId}/classes/{classId}/teachers": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getClassTeachers",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "classId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimetableTeacherArray"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/timetable/{timetableId}/teachers/": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getAllTeachers",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "query",
            "description": "The name of the teaacher to look for",
            "required": false,
            "type": "string"
          },
          {
            "name": "color",
            "in": "query",
            "description": "The color of the teacher to look for",
            "required": false,
            "type": "string"
          },
          {
            "name": "taught_classes",
            "in": "query",
            "description": "An array of classes the teacher should be teaching",
            "required": false,
            "type": "array",
            "schema": {
              "type": "string",
              "example": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimetableTeacherArray"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/timetable/{timetableId}/teachers/{teacherId}": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getTeacherById",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "teacherId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimetableTeacher"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/timetable/{timetableId}/teachers/{teacherId}/taught_classes": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getTeacherTaughtClasses",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "teacherId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/timetable/{timetableId}/subjects/": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getAllSubjects",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "query",
            "description": "The name of the subject to look for",
            "required": false,
            "type": "string"
          },
          {
            "name": "short_name",
            "in": "query",
            "description": "The short name of the subject to look for",
            "required": false,
            "type": "string"
          },
          {
            "name": "color",
            "in": "query",
            "description": "The color of the subject to look for",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimetableSubjectArray"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/timetable/{timetableId}/subjects/{subjectId}": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getSubjectById",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "subjectId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimetableSubject"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/timetable/{timetableId}/lessons/{lessonId}/teachers": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getLessonTeachers",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimetableTeacherArray"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/timetable/{timetableId}/lessons/{lessonId}/subject": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getLessonSubject",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimetableSubject"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "TimetableTeacher": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": ""
        },
        "name": {
          "type": "string",
          "example": ""
        },
        "color": {
          "type": "string",
          "example": ""
        },
        "taught_classes": {
          "type": "array",
          "example": [
            ""
          ],
          "items": {
            "type": "string"
          }
        }
      }
    },
    "TimetableTeacherArray": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": ""
          },
          "name": {
            "type": "string",
            "example": ""
          },
          "color": {
            "type": "string",
            "example": ""
          },
          "taught_classes": {
            "type": "array",
            "example": [
              ""
            ],
            "items": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "TimetableSubject": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": ""
        },
        "name": {
          "type": "string",
          "example": ""
        },
        "short_name": {
          "type": "string",
          "example": ""
        },
        "color": {
          "type": "string",
          "example": ""
        }
      }
    },
    "TimetableSubjectArray": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": ""
          },
          "name": {
            "type": "string",
            "example": ""
          },
          "short_name": {
            "type": "string",
            "example": ""
          },
          "color": {
            "type": "string",
            "example": ""
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}